Transport: make the caller acknowledge the onDisconnected event.
Since the onDisconnected callback is invoked from a netty thread,
it's easy for the caller to get reconnected without realizing that
it has been disconnected.

With this patch, the transport stops sending/receiving messages
to/from the destination once the onDisconnected callback is invoked.
To reestablish the connection, the caller must explicitly call the
clear() method.